Split router into cohesive modules; drop the bare-fragment vestige
Reorganize the 664-line router.ts into focused modules so the package
structure reads cleanly:
- types.ts  — shared interfaces (no logic)
- seams.ts  — window.* bridges to the optional client runtime
- cache.ts  — snapshot cache (stale-while-revalidate + LRU)
- outlet.ts — read the response: extract outlet + island modules
- router.ts — the controller wiring those together

Remove X-Barefoot-Navigate and the bare-fragment response path entirely:
the router always fetches the full page and extracts [bf-outlet] client-
side, so extractOutlet now returns null when the marker is absent (the
caller hard-navigates) with no full-document heuristic. Drop the header
assertion and the bare-fragment test, and the related README/doc-comment
prose. Reframe the README from "prototype" to "alpha".

No behavior change for full-page navigation; tests stay green (17).
