+++ packages/router/src/index.ts
@@ -1,5 +1,5 @@
 export { startRouter, navigate } from './router.ts'
-export type { RouterOptions, NavigateOptions, Router } from './router.ts'
+export type { RouterOptions, NavigateOptions, Router } from './types.ts'
 
 // Re-exported for server-side helpers that mark the swappable region.
 export { BF_OUTLET } from '@barefootjs/shared'
